*** empty log message ***

+++ Macro_Machine.h
@@ -30,7 +30,7 @@ typedef struct
 
   int* symbol;
 
-  int                total_symbols;
+  unsigned long long total_symbols;
   unsigned long long total_steps;
 
   int position;
